sqlite,persist: add generic tests for both stores
